2671 Fix open file bottom sheet for tv
