+++ app/src/main/res/drawable/popup_black.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_pressed="true">
+        <color
+            android:color="@color/colorPrimary"/>
+    </item>
+    <item>
+        <color
+            android:color="#655611"/>
+    </item>
+</selector>
